requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
earthengine-api==0.1.365
geemap==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
joblib==1.3.1
python-dotenv==1.0.0
gunicorn==20.1.0
geopandas==0.14.0
shapely==2.0.1
requests==2.31.0
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
